+++ tests/utils/CMakeLists.txt
@@ -1,9 +1,36 @@
-add_library(RuntimeTestUtils
-            RuntimeTestUtils.cpp)
-target_link_libraries(RuntimeTestUtils
+add_library(ExecutorTestBuilder
+            ExecutorTestBuilder.cpp)
+target_link_libraries(ExecutorTestBuilder
+                      PRIVATE
+                        benchmark
+                        Base
+                        DeviceManager
+                        ExecutorUnitTestWrapper
+                        Graph
+                        TestDeviceManager)
+
+add_library(ExecutorBenchmarkWrapper
+            ExecutorBenchmarkWrapper.cpp)
+target_link_libraries(ExecutorBenchmarkWrapper
+                      PRIVATE
+                        benchmark
+                        Base
+                        Executor
+                        Graph)
+
+add_library(ExecutorUnitTestWrapper
+            ExecutorUnitTestWrapper.cpp)
+target_link_libraries(ExecutorUnitTestWrapper
+                      PRIVATE
+                        Base
+                        Executor
+                        Graph)
+
+add_library(TestDeviceManager
+            TestDeviceManager.cpp)
+target_link_libraries(TestDeviceManager
                       PRIVATE
                         Base
                         DeviceManager
-                        Executor
                         Graph
                         ThreadPool)
