+++ tests/utils/ExecutorBenchmarkWrapper.h
@@ -1,0 +1,50 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#ifndef GLOW_TESTS_UTILS_EXECUTORBENCHMARKWRAPPER_H
+#define GLOW_TESTS_UTILS_EXECUTORBENCHMARKWRAPPER_H
+
+#include "ExecutorTestWrapper.h"
+
+#include "benchmark/benchmark.h"
+
+namespace glow {
+namespace runtime {
+
+class ExecutorBenchmarkWrapper final : public ExecutorTestWrapper {
+public:
+  using PlaceholderNameMapTy = ExecutorTestWrapper::PlaceholderNameMapTy;
+  using DAGNodeNameMapTy = ExecutorTestWrapper::DAGNodeNameMapTy;
+
+  /// Constructor.
+  ExecutorBenchmarkWrapper(const std::shared_ptr<Executor> &executor,
+                           std::unique_ptr<DAGNode> root,
+                           std::unique_ptr<Type> type, DAGNodeNameMapTy nodes,
+                           PlaceholderNameMapTy placeholders,
+                           std::unique_ptr<ExecutionContext> inputContext,
+                           std::unique_ptr<ExecutionContext> outputContext,
+                           RunIdentifierTy runId, bool expectSuccess)
+      : ExecutorTestWrapper(executor, std::move(root), std::move(type),
+                            std::move(nodes), std::move(placeholders),
+                            std::move(inputContext), std::move(outputContext),
+                            runId, expectSuccess) {}
+
+  /// Run the benchmark.
+  bool run(benchmark::State &state);
+};
+} // namespace runtime
+} // namespace glow
+
+#endif // GLOW_TESTS_UTILS_EXECUTORBENCHMARKWRAPPER_H
